Refactor NotFound components into a centralized implementation
Consolidated the NotFound logic into a single, reusable component under 'not-found'. Removed redundant code and simplified its usage across components, improving maintainability and consistency.
